x2gokdrive{remote,selection}.c: wrap initializer into another set of curly braces until the first POD is reached.
Fixes compile errors on EPEL 7 and generally systems with older GCC
versions.

That warning, which is turned into an error later on, is generally
overly strict, since {0} is widely used as an "universal zero
initializer" in C, initializing all other elements that are not
explicitly given in the list implicitly with zero.

For some reason, GCC decided that it is an error to initialize a whole
struct that way and explicitly requires that at least one plain-old-data
type is initialized that way.

Case in point, newer GCC versions removed that warning altogether and
just blindly accepts the usual syntax.