debian/control: make libxfont1-dev an entirely optional dependency and libxfont-dev a hard one.
This should always install libxfont-dev (which typically is the
libXfont2-version), but also the libXfont1-version if available.

Ubuntu 16.04/xenial has X.Org Server 1.18.4 which is not compatible with
libXfont2, but Debian 9/stretch uses X.Org Server 1.19.2, which is
compatible with libXfont2 (and actually requires it), but also has both
packages available.